Pre-Alpha 4.0.7 Questing System!
Starting working on a brand new questing system consisting of Quests, Events, and Paths. Fairly simple baseline design that leaves a world of expansion and possibility to build on. Thanks to Holistic 3D on youtube for the tutorial. Still in progress. Made some edits to the blender file in terms of animations