Pre-Alpha 4.0.13 Quest system work continues
Added quest failed prefabs, as well as add a new quest type, checking for the amount of an inventory item. (Currently not working properly). Lots of tiny bug fixes to the inventory system, including that annoying bug I couldn't fix a while back. The inventory not showing names & numbers occasionally is still not fixed and idk what is causing it at this point.