Pre-Alpha 4.1.1 Re-writing Charcontrol + Universal Simple Timer
Decided the old code needed to be brought up to speed. I was a fool to think it would be easy. This is easiest the most grueling thing I've done in this god forsaken project, but rewriting from scratch would take too long. May the gods give me strength (and motivation to get through it). Charcontrol code temporarily moved to Charcontrol_wastebin to trim off excess. Also made a timer script that can be called from anywhere to make a neat timer, how convenient.