Pre-Alpha 4.1.3 Walking and Rolling Animations + Teleporters Overhaul
Walking and Rolling animations have successfully been implemented and work well. Animations themselves could use some polish from in blender but I cannot be bothered. The walking loop is slightly too long and causes some weird looking "extra walking" when transitioning back to idle. Other than that, works great.
Changed the teleporter menu to a sliding one instead of pages, and future proofed it so that you can add an infinite amount of teleporters and it still works. That annoying UI bug where the last elements selected with a keyboard not working still appears. Need to find a fix.