Pre-Alpha 4.1.16 More general animation stuff + Brand new Credits!
Added a few animations, yet to be implemented (definitely not putting that off) and added proper scrolling credits! Automatically calculates how far to scroll based on its height, and can be changed on the fly. The code is held together with duct tape and 2x4s, so hopefully I future proofed enough that I don't have to touch it, if not, you're welcome future me :)