Pre-Alpha 4.2.0 Combo Code is IN
Worked the whole day on the single function that checks to make sure combos are executed properly. After HOURS of trial and error, I finally got it to work (and it works perfectly). Combos are defined by an Attack List. Combos and Attacks are their own classes that hold their own information. The method checks the attack inputted against the attacks in the combos. Now to blend it with the animations....whew