Coming along great, almost done refactoring and integrating dialogue.
Added alot of functionality to the QuestGiver and QuestManager scripts. Also used the [ButtonMethod] attribute from Mybox to give the quest cleanup script some usage as a debug tool. No need for debug spheres anymore! Currently stuck on an issue where I'm trying to get quest events to update the default tree ID in dialogue, but it happens repeatedly when it shouldn't. Should be ready to do a PR soon