Pre-Alpha 4.4.1 Info Hub works FLAWLESSLY
You can add as many, or as few pages as you want, and the code will switch between them properly, you can also click the buttons at the top to go straight to its menu. Still need to add functionality for buttons to be highlight for their page, but other than that it's done.