Pre-Alpha 4.4.5 Quest IHUI is working well, coding up next.
Scrapped the old UI idea as it wasn't working and came up with a new, page version. No getting all the UI elements to work via their code is the newest challenge. Shouldn't be too hard since I can reuse a lot of the code from the IHUI pages code.