Pre-Alpha 4.4.9 The IHUI Quest Page is complete! (ish)
Everything works as expected, however there are numerous bugs. Not focusing on that currently, as I feel like I need something I can show. This is great, and needed, but its not the main focus of what I want to show in a trailer or demo. Gonna move onto something more visual and a better showcase of my work.