Pre-Alpha 4.6.13 Combat is getting out of hand
Currently stuck with a simple, but huge bug. How do I get a function to trigger reliably at the end of every combat animation without using animation events or animstate scripts? Currently trying a coroutine that runs for the same amount of time as the animation, but it does not work. While I would love to continue working on this today, it would probably take up the rest of my Sunday, which isnt wise. Good luck future me.