Pre-Alpha 4.6.15 Combos FINALLY WORK PROPERLY. NOTE: Error in this commit
By using state behaviors and some serious bug fixing, finally got this thing to work with a simple light combo. Now to make sure it works with other combos and various other things. NOTE: This commit contains a bug in unity that didn't serialize the Player GUID properly, so its missing, or causes Unity to crash. Attempting to fix.