Pre Alpha 4.12.0 UI Toolkit work begins!
Started using UI Toolkit in an effort to never have to use Unity's terrible old UI system. It uses XML and USS which is remarkably similar to HTML and CSS, which is great as I wanted to learn them. Began with the new inventory screen, it's going well so far...