Pre-Alpha 4.13.0 Added some Dialogue things and made OffPaths
Added dialogue sounds, reworked some things and gave up on the multiple characters in a scene.

Added OffPaths. These are doors that take the player "outside the 2d plane" to rooms or another path. The player transitions through them by fading in and out. Need to add a fade to black or wipe transition.